relocate some modules, packages, and classes
