+++ vlog4j-core/src/main/java/org/semanticweb/vlog4j/core/model/api/PrefixDeclarationException.java
@@ -1,8 +1,8 @@
-package org.semanticweb.vlog4j.syntax.parser;
+package org.semanticweb.vlog4j.core.model.api;
 
 /*-
  * #%L
- * vlog4j-syntax
+ * vlog4j-parser
  * %%
  * Copyright (C) 2018 - 2019 VLog4j Developers
  * %%
@@ -20,31 +20,13 @@ package org.semanticweb.vlog4j.syntax.parser;
  * #L%
  */
 
-public class ParsingException extends Exception {
+public class PrefixDeclarationException extends Exception {
+    /** 
+     * 
+     */
+    private static final long serialVersionUID = 1L; 
 
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 2849123381757026724L;
-
-	public ParsingException() {
-		super();
-	}
-
-	public ParsingException(String message) {
-		super(message);
-	}
-
-	public ParsingException(Throwable cause) {
-		super(cause);
-	}
-
-	public ParsingException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-	public ParsingException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-		super(message, cause, enableSuppression, writableStackTrace);
-	}
-
+    public PrefixDeclarationException(String errorMessage) {
+        super(errorMessage);
+    }   
 }
