+++ vlog4j-examples/src/main/java/org/semanticweb/vlog4j/examples/ExamplesUtils.java
@@ -37,6 +37,8 @@ import org.semanticweb.vlog4j.core.model.implementation.Expressions;
 import org.semanticweb.vlog4j.core.reasoner.Reasoner;
 import org.semanticweb.vlog4j.core.reasoner.exceptions.ReasonerStateException;
 import org.semanticweb.vlog4j.core.reasoner.implementation.QueryResultIterator;
+import org.semanticweb.vlog4j.parser.ParsingException;
+import org.semanticweb.vlog4j.parser.RuleParser;
 
 public final class ExamplesUtils {
 
@@ -82,12 +84,9 @@ public final class ExamplesUtils {
 	 * Prints out the answers given by {@code reasoner} to the query
 	 * ({@code queryAtom}).
 	 *
-	 * @param queryAtom
-	 *            query to be answered
-	 * @param reasoner
-	 *            reasoner to query on
-	 * @throws ReasonerStateException
-	 *             in case the reasoner has not yet been loaded.
+	 * @param queryAtom query to be answered
+	 * @param reasoner  reasoner to query on
+	 * @throws ReasonerStateException in case the reasoner has not yet been loaded.
 	 */
 	public static void printOutQueryAnswers(final PositiveLiteral queryAtom, final Reasoner reasoner)
 			throws ReasonerStateException {
@@ -99,12 +98,30 @@ public final class ExamplesUtils {
 	}
 
 	/**
+	 * Prints out the answers given by {@code reasoner} to the query
+	 * ({@code queryAtom}).
+	 *
+	 * @param queryAtom query to be answered
+	 * @param reasoner  reasoner to query on
+	 * @throws ReasonerStateException in case the reasoner has not yet been loaded.
+	 */
+	public static void printOutQueryAnswers(final String queryString, final Reasoner reasoner)
+			throws ReasonerStateException {
+		RuleParser ruleParser = new RuleParser();
+		try {
+			PositiveLiteral query = ruleParser.parsePositiveLiteral(queryString);
+			printOutQueryAnswers(query, reasoner);
+		} catch (ParsingException e) {
+			throw new RuntimeException(e.getMessage(), e);
+		}
+	}
+
+	/**
 	 * Returns the size of an iterator.
 	 *
 	 * @FIXME This is an inefficient way of counting results. It should be done at a
 	 *        lower level instead
-	 * @param Iterator<T>
-	 *            to iterate over
+	 * @param Iterator<T> to iterate over
 	 * @return number of elements in iterator
 	 */
 	public static <T> int iteratorSize(Iterator<T> iterator) {
@@ -117,10 +134,8 @@ public final class ExamplesUtils {
 	/**
 	 * Creates an Atom with @numberOfVariables distinct variables
 	 *
-	 * @param predicateName
-	 *            for the new predicate
-	 * @param arity
-	 *            number of variables
+	 * @param predicateName for the new predicate
+	 * @param arity         number of variables
 	 */
 	private static PositiveLiteral makeQueryAtom(String predicateName, int arity) {
 		final List<Term> vars = new ArrayList<>();
@@ -132,12 +147,9 @@ public final class ExamplesUtils {
 	/**
 	 * Exports the extension of the Atom with name @predicateName
 	 *
-	 * @param reasoner
-	 *            reasoner to query on
-	 * @param atomName
-	 *            atom's name
-	 * @param arity
-	 *            atom's arity
+	 * @param reasoner reasoner to query on
+	 * @param atomName atom's name
+	 * @param arity    atom's arity
 	 */
 	public static void exportQueryAnswersToCSV(Reasoner reasoner, String atomName, int arity)
 			throws ReasonerStateException, IOException {
