+++ vlog4j-parser/pom.xml
@@ -11,9 +11,9 @@
 		<version>0.4.0-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>vlog4j-syntax</artifactId>
+	<artifactId>vlog4j-parser</artifactId>
 
-	<name>VLog4j Syntax</name>
+	<name>VLog4j Parser</name>
 	<url>http://maven.apache.org</url>
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
