+++ vlog4j-parser/src/main/java/org/semanticweb/vlog4j/parser/RuleParser.java
@@ -1,4 +1,4 @@
-package org.semanticweb.vlog4j.syntax.parser;
+package org.semanticweb.vlog4j.parser;
 
 /*-
  * #%L
@@ -22,17 +22,16 @@ package org.semanticweb.vlog4j.syntax.parser;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
-import java.time.format.FormatStyle;
 import java.util.List;
 
 import org.semanticweb.vlog4j.core.model.api.Literal;
 import org.semanticweb.vlog4j.core.model.api.PositiveLiteral;
+import org.semanticweb.vlog4j.core.model.api.PrefixDeclarationException;
 import org.semanticweb.vlog4j.core.model.api.Rule;
-import org.semanticweb.vlog4j.parser.implementation.javacc.JavaCCRuleParser;
-import org.semanticweb.vlog4j.parser.implementation.javacc.ParseException;
-import org.semanticweb.vlog4j.parser.implementation.javacc.TokenMgrError;
-import org.semanticweb.vlog4j.syntax.common.PrefixDeclarationException;
-import org.semanticweb.vlog4j.syntax.parser.RuleParserBase.FormulaContext;
+import org.semanticweb.vlog4j.parser.RuleParserBase.FormulaContext;
+import org.semanticweb.vlog4j.parser.javacc.JavaCCRuleParser;
+import org.semanticweb.vlog4j.parser.javacc.ParseException;
+import org.semanticweb.vlog4j.parser.javacc.TokenMgrError;
 
 /**
  * Class to access VLog parsing functionality.
