+++ vlog4j-parser/src/main/java/org/semanticweb/vlog4j/parser/javacc/.gitignore
